Make sure that array indices are integer
Was getting deprecation warnings, such as the one below, because
of float values being passed as indices to the array
/home/edward/lib/Ornithokrites/features.py:75: DeprecationWarning: using
a non-integer number instead of an integer will result in an error in
the future
  self.ExtractedFeatures[i,10] = SpectralRolloff[start:end].min()

+++ features.py
@@ -59,8 +59,8 @@ class FeatureExtractor(object):
         SpectralRolloff = self.Features['SpectralRolloff']
         
         for i, segment in enumerate(segments):
-            start = segment[0] / frame_overlap - 1
-            end = segment[1] / frame_overlap + 1
+            start = int(segment[0] / frame_overlap) - 1
+            end = int(segment[1] / frame_overlap) + 1
             
             self.ExtractedFeatures[i,0] = LPC1[start:end].mean()
             self.ExtractedFeatures[i,1] = LSF7[start:end].min()
